requirements.txt
dotenv
faiss-cpu
google
itemadapter
langchain
langchain_huggingface
numpy==1.26.4
scrapy
fastapi
pydantic
uvicorn
google-generativeai
langchain-community
huggingface_hub[hf_xet]
python-dotenv
aiohttp
certifi
supabase
httpx
transformers
torch